Make reformatMetadata a core function
Use reformatMetadata in CheckedContract

Remove the settings flag in validation script
