+++ services/validation/test/expected-output/single.json
@@ -1,1 +1,1 @@
-{"language":"Solidity","sources":{"browser/1_Storage.sol":{"keccak256":"0xaedc7086ad8503907209f50bac1e4dc6c2eca2ed41b15d03740fea748ea3f88e","urls":["bzz-raw://4bc331951c25951321cb29abbd689eb3af669530222c6bb2d45ff45334ee83a7","dweb:/ipfs/QmWb1NQ6Pw8ZLMFX8uDjMyftgcEieT9iP2TvWisPhjN3U2"],"content":"pragma solidity >=0.4.22 <0.7.0;\n\n/**\n * @title Storage\n * @dev Store & retreive value in a variable\n */\ncontract Storage {\n\n    uint256 number;\n\n    /**\n     * @dev Store value in variable\n     * @param num value to store\n     */\n    function store(uint256 num) public {\n        number = num;\n    }\n\n    /**\n     * @dev Return value \n     * @return value of 'number'\n     */\n    function retreive() public view returns (uint256){\n        return number;\n    }\n}"}},"settings":{"outputSelection":{"*":{"*":["evm.bytecode","abi"]}},"remappings":[],"optimizer":{"enabled":false,"runs":200},"evmVersion":"istanbul","metadata":{"bytecodeHash":"ipfs"},"libraries":{}}}
+{"settings":{"evmVersion":"istanbul","libraries":{"":{}},"metadata":{"bytecodeHash":"ipfs"},"optimizer":{"enabled":false,"runs":200},"remappings":[],"outputSelection":{"browser/1_Storage.sol":{"Storage":["evm.bytecode","evm.deployedBytecode","metadata"]}}},"sources":{"browser/1_Storage.sol":{"content":"pragma solidity >=0.4.22 <0.7.0;\n\n/**\n * @title Storage\n * @dev Store & retreive value in a variable\n */\ncontract Storage {\n\n    uint256 number;\n\n    /**\n     * @dev Store value in variable\n     * @param num value to store\n     */\n    function store(uint256 num) public {\n        number = num;\n    }\n\n    /**\n     * @dev Return value \n     * @return value of 'number'\n     */\n    function retreive() public view returns (uint256){\n        return number;\n    }\n}"}},"language":"Solidity"}
